backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
resend==0.8.0
# Security dependencies
fastapi-users[sqlalchemy]==12.1.2
slowapi==0.1.9
bcrypt==4.0.1
python-frontmatter
